backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rodney Souza <ro.rss.129@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.0"}
uvicorn = "^0.30.6"
motor = {extras = ["srv"], version = "^3.6.0"}
pymongo = "^4.9.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
pydantic = {extras = ["dotenv"], version = "^2.9.2"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
